Install server dependencies on a droplet

Installs R and common dependencies needed for running _server.yml applications. This is called automatically by do_provision() but can be called separately if needed.

Usage

do_install_server_deps(droplet, keyfile = do_keyfile())

Arguments

droplet

The DigitalOcean droplet that you want to provision (see analogsea::droplet()). If empty, a new DigitalOcean server will be created.

keyfile

Path to private key for authentication. By default, uses the key for "digitalocean.com" from ssh::ssh_key_info().

Value

Invisibly returns NULL. Called for side effects.

Examples

if (FALSE) { # \dontrun{
  # Reinstall or update server dependencies on an existing droplet
  droplet <- analogsea::droplet(123456)
  do_install_server_deps(droplet)
} # }